Address PR feedback: improve vector growth model, remove duplicates, and standardize types
- Update ExponentialWithVectorGrowthModel to reflect Neo4j vector index formula
  and assumption that chunk size + vector >> entity size for RAG workloads
  - Increase default vector_proportion from 0.3 to 0.7 (70%)
  - Add documentation referencing Neo4j vector index size calculation formula
  - Simplify implementation to single code path

- Remove duplicate SizingCalculations model, use Neo4jSizingCalculationResult directly
  - Eliminates unnecessary conversion step in service layer
  - Simplifies codebase by removing redundant model

- Remove unnecessary backward compatibility alias
  - Remove 'SizingCalculator' alias from __init__.py as this is first release

- Improve type hints and validation
  - Change memory_to_storage_ratio from float to int (only accepts 1, 2, 4, 8)
  - Update default values from 1.0 to 1
  - Add clear documentation that ValueError is raised for invalid values
  - Update all type hints across service, calculator, projector, and API layers

- Remove unnecessary validation check
  - Remove annual_growth_rate > 1000 check (unhelpful error message)

- Simplify imports
  - Use direct import for Neo4jSizingCalculationResult instead of TYPE_CHECKING
  (no circular import exists)

+++ servers/mcp-neo4j-cloud-aura-api/src/mcp_neo4j_aura_manager/aura_manager.py
@@ -156,7 +156,7 @@ class AuraManager:
         percentage_nodes_with_vector_properties: Optional[float] = None,
         number_of_vector_indexes: Optional[int] = None,
         quantization_enabled: bool = False,
-        memory_to_storage_ratio: Optional[float] = None,
+        memory_to_storage_ratio: Optional[int] = None,
         concurrent_end_users: Optional[int] = None,
         **kwargs
     ) -> Dict[str, Any]:
@@ -191,7 +191,7 @@ class AuraManager:
         projection_years: int = 3,
         workloads: Optional[List[str]] = None,
         domain: Optional[str] = None,
-        memory_to_storage_ratio: float = 1.0,
+        memory_to_storage_ratio: int = 1,
         **kwargs
     ) -> Dict[str, Any]:
         """Forecast database growth over multiple years."""
